+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Change Log
 
-### 1.3.x
+### 1.3.6
 
 * Fixed:    Pull ups of columns now respect types when filtering in MySql
+* Added:    Support for aggregation of pull ups.
 
 ### 1.3.5